requirements.txt
numpy
pandas
pyarrow
scikit-learn
onnx
onnxruntime
pyyaml
joblib
skl2onnx
matplotlib
streamlit
plotly
